% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rscimark.R
\name{rscimark}
\alias{rscimark}
\title{rscimark}
\usage{
rscimark(large = FALSE, minimum.time = 2)
}
\arguments{
\item{large}{[logical(1)]\cr
Run large version of benchmark?
Default is \code{FALSE}.}

\item{minimum.time}{[numeric(1)]\cr
Minimum time to run each of the benchmarks, in seconds.
Default is 2.}
}
\description{
fdd
}

